Cast fileInfo’s uncompressed_size to unsigned to avoid Xcode warning
This warning didn’t appear when I first submitted the pull request, but it does now. This cast avoids it.

+++ SSZipArchive/SSZipArchive.m
@@ -948,7 +948,7 @@ BOOL _fileIsSymbolicLink(const unz_file_info *fileInfo);
             // Is this the file we are looking for?
             if ([strPath isEqualToString:name]) {
                 data = [[NSMutableData alloc] initWithLength:fileInfo.uncompressed_size];
-                int readBytes = unzReadCurrentFile(zip, data.mutableBytes, fileInfo.uncompressed_size);
+                int readBytes = unzReadCurrentFile(zip, data.mutableBytes, (unsigned)fileInfo.uncompressed_size);
                 if (readBytes != fileInfo.uncompressed_size) {
                     NSDictionary *userInfo = @{NSLocalizedDescriptionKey: @"Failed to read contents of file entity"};
                     unzippingError = [NSError errorWithDomain:SSZipArchiveErrorDomain
